# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ConversationMode = typeof ConversationMode[keyof typeof ConversationMode];


export const ConversationMode = {
  listen: 'listen',
  understand: 'understand',
  reframe: 'reframe',
  help: 'help',
  private: 'private',
} as const;

export interface Conversation {
  id: number;
  title: string;
  mode: ConversationMode;
  createdAt: string;
  updatedAt: string;
}

export type ConversationInputMode = typeof ConversationInputMode[keyof typeof ConversationInputMode];


export const ConversationInputMode = {
  listen: 'listen',
  understand: 'understand',
  reframe: 'reframe',
  help: 'help',
  private: 'private',
} as const;

export interface ConversationInput {
  title?: string;
  mode?: ConversationInputMode;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
  /** @nullable */
  emotion?: string | null;
}

export type MessageInputMode = typeof MessageInputMode[keyof typeof MessageInputMode];


export const MessageInputMode = {
  listen: 'listen',
  understand: 'understand',
  reframe: 'reframe',
  help: 'help',
  private: 'private',
} as const;

export interface MessageInput {
  /** @minLength 1 */
  content: string;
  mode?: MessageInputMode;
}

export interface EmotionDetectionInput {
  /** @minLength 1 */
  content: string;
}

export interface EmotionDetection {
  primary: string;
  secondary: string[];
  intensity: number;
  signals: string[];
  reflection: string;
}

export type JournalEntryEntryType = typeof JournalEntryEntryType[keyof typeof JournalEntryEntryType];


export const JournalEntryEntryType = {
  open: 'open',
  guided: 'guided',
} as const;

export interface JournalEntry {
  id: number;
  title?: string;
  content: string;
  mood?: string;
  /** @nullable */
  moodTag?: string | null;
  entryType?: JournalEntryEntryType;
  createdAt: string;
  updatedAt?: string;
}

export type JournalEntryInputEntryType = typeof JournalEntryInputEntryType[keyof typeof JournalEntryInputEntryType];


export const JournalEntryInputEntryType = {
  open: 'open',
  guided: 'guided',
} as const;

export interface JournalEntryInput {
  title?: string;
  /** @minLength 1 */
  content: string;
  mood?: string;
  moodTag?: string;
  entryType?: JournalEntryInputEntryType;
}

export type JournalEntryUpdateEntryType = typeof JournalEntryUpdateEntryType[keyof typeof JournalEntryUpdateEntryType];


export const JournalEntryUpdateEntryType = {
  open: 'open',
  guided: 'guided',
} as const;

export interface JournalEntryUpdate {
  title?: string;
  /** @minLength 1 */
  content?: string;
  mood?: string;
  moodTag?: string;
  entryType?: JournalEntryUpdateEntryType;
}

export interface Memory {
  id: number;
  label: string;
  detail: string;
  enabled: boolean;
  createdAt: string;
}

export interface MemorySettings {
  enabled: boolean;
}

export interface EmotionCount {
  emotion: string;
  count: number;
  color: string;
}

export interface IntensityPoint {
  day: string;
  value: number;
}

export interface DashboardSummary {
  checkIns: number;
  journalEntries: number;
  conversations: number;
  streak: number;
  topEmotions: EmotionCount[];
  weeklyIntensity: IntensityPoint[];
}

export interface CompanionBootstrap {
  conversations: Conversation[];
  messages: Message[];
  memories: Memory[];
  memoryEnabled: boolean;
  dashboard: DashboardSummary;
  journalEntries: JournalEntry[];
}

export interface PrivateNote {
  id: number;
  content: string;
  isLetter: boolean;
  createdAt: string;
  /** @nullable */
  expiresAt?: string | null;
}

export interface PrivateNoteInput {
  /** @minLength 1 */
  content: string;
  isLetter?: boolean;
  expiresAt?: string;
}