# lib/api-zod/src/generated/types/journalEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JournalEntryEntryType } from './journalEntryEntryType';

export interface JournalEntry {
  id: number;
  title?: string;
  content: string;
  mood?: string;
  /** @nullable */
  moodTag?: string | null;
  entryType?: JournalEntryEntryType;
  createdAt: string;
  updatedAt?: string;
}