# lib/api-zod/src/generated/types/journalEntryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JournalEntryInputEntryType } from './journalEntryInputEntryType';

export interface JournalEntryInput {
  title?: string;
  /** @minLength 1 */
  content: string;
  mood?: string;
  moodTag?: string;
  entryType?: JournalEntryInputEntryType;
}