# lib/api-zod/src/generated/types/privateNote.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PrivateNote {
  id: number;
  content: string;
  isLetter: boolean;
  createdAt: string;
  /** @nullable */
  expiresAt?: string | null;
}